requirements.txt
tqdm
pytorch-ignite
torchsummary